BUG: rbf: don't instantiate rbf function at parse time
By default, we now can specify a str arg for RBFInt(..., rbf=...) in
addition to the old API (pass a RBFFunction instance). That will make
sure that we always create an instance of RBFFunction at RBFInt
instantiation, not already at code compile/parse time. The old behavior
leads to subtle bugs where the same RBFFunction instance attached to
each created RBFInt instance is shared, if the `rbf` kwd is not
explicitely set.