MNT: fix numpy deprecation warning
Use tuples instead of lists for passing slices for multiple dimensions:
    arr[(slice(...),...)]   # ok
    arr[[slice(...),...]]   # nope

Solves:

FutureWarning: Using a non-tuple sequence for multidimensional indexing
is deprecated; use `arr[tuple(seq)]` instead of `arr[seq]`. In the
future this will be interpreted as an array index, `arr[np.array(seq)]`,
which will result either in an error or a different result.