MNT: mpl: handle matplotlib deprecation
Use plt.subplots() instead of the old
    fig = plt.figure()
    ax = fig.add_subplot(111)

Solves:

MatplotlibDeprecationWarning: Adding an axes using the same arguments as
a previous axes currently reuses the earlier instance.  In a future
version, a new instance will always be created and returned.  Meanwhile,
this warning can be suppressed, and the future behavior ensured, by
passing a unique label to each axes instance.
  ax = fig.add_subplot(111)

However, matplotlib doesn't complain when we use the old-ish style in
fig_ax3d() so don't change that.