MNT: docs: require sphinx < 3.4
As of sphinx 3.4, we get

    $ cd pwtools/doc
    $ sphinx-autodoc ...
    $ make html
    Exception occurred:
      File "/usr/lib/python3/dist-packages/sphinx/pycode/__init__.py", line 68, in get_module_source
        raise PycodeError('source is not a .py file: %r' % filename)
    sphinx.errors.PycodeError: source is not a .py file: '/home/elcorto/soft/git/pwtools/pwtools/_dcd.so'

when using

    [ source/generated/api/__sphinx_autodoc_module__pwtools._dcd.rst ]

    .. automodule:: pwtools._dcd
       :no-members:
       :no-inherited-members:

    .. currentmodule:: pwtools._dcd

    .. Links to members.
    .. autosummary::
       :toctree:

       get_dcd_file_info
       open_file
       read_dcd_data
       read_dcd_data_from_unit
       read_dcd_header_from_unit

The error is triggered when sphinx starts processing the module members
listed in the autosummary directive. Not listing module members in
autosummary works, but then of course we don't have their docs.

Older Sphinx versions just read e.g.
pwtools._dcd.get_dcd_file_info.__doc__ but now it thinks the module is a
py file and wants to open it. Need to file a bug report against Sphinx
but for now exclude 3.4.x .