TST: prefer executable pytest over pytest-3
When in a venv, make sure we don't use the system's pytest b/c that
often leads to problems (e.g. system pytest cannot find the package
under test within a venv).

In a venv, use pytest, which is the executable name installed when `pip
install pytest`, whereas `pytest-3` is usually a system version such as
`/usr/bin/pytest-3` that we wish to use when not in a venv.