ENH: rbf.hyperopt.fit_opt() more sanity checks
Catch passing p and/or r in rbf_kwds when tying to optimize those. These
would cause an exception down in FitError anyway, but now we catch them
here and a add a better error message. Also remove the part where we
remove those from rbf_kwds before building the final model b/c if
rbf_kwds was passed in the correct format, here was actually nothing to
remove.